# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DebateSphere API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  USER: 'USER',
  ADMIN: 'ADMIN',
} as const;

export type UserTier = typeof UserTier[keyof typeof UserTier];


export const UserTier = {
  Bronze_Debater: 'Bronze Debater',
  Silver_Debater: 'Silver Debater',
  Gold_Debater: 'Gold Debater',
  Diamond_Debater: 'Diamond Debater',
} as const;

export interface User {
  id: number;
  username: string;
  email: string;
  role: UserRole;
  reputation: number;
  tier: UserTier;
  /** @nullable */
  bio?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserProfile {
  id: number;
  username: string;
  role: string;
  reputation: number;
  tier: string;
  /** @nullable */
  bio?: string | null;
  debateCount: number;
  argumentCount: number;
  totalVotesReceived: number;
  createdAt: string;
}

export type DebateStatus = typeof DebateStatus[keyof typeof DebateStatus];


export const DebateStatus = {
  active: 'active',
  closed: 'closed',
  archived: 'archived',
} as const;

export interface Debate {
  id: number;
  title: string;
  description: string;
  category: string;
  tags: string[];
  status: DebateStatus;
  authorId: number;
  author: User;
  argumentCount: number;
  participantCount: number;
  viewCount?: number;
  createdAt: string;
  updatedAt: string;
}

export interface DebateList {
  debates: Debate[];
  total: number;
}

export interface DebateInput {
  /** @minLength 10 */
  title: string;
  /** @minLength 20 */
  description: string;
  category: string;
  tags?: string[];
}

export type DebateUpdateStatus = typeof DebateUpdateStatus[keyof typeof DebateUpdateStatus];


export const DebateUpdateStatus = {
  active: 'active',
  closed: 'closed',
  archived: 'archived',
} as const;

export interface DebateUpdate {
  title?: string;
  description?: string;
  category?: string;
  tags?: string[];
  status?: DebateUpdateStatus;
}

export interface Category {
  name: string;
  count: number;
}

export interface JoinResponse {
  debateId: number;
  participantCount: number;
}

export type ArgumentStance = typeof ArgumentStance[keyof typeof ArgumentStance];


export const ArgumentStance = {
  for: 'for',
  against: 'against',
  neutral: 'neutral',
} as const;

export interface Argument {
  id: number;
  content: string;
  debateId: number;
  authorId: number;
  author: User;
  upvotes: number;
  downvotes: number;
  replyCount: number;
  stance: ArgumentStance;
  /** @nullable */
  userVote?: string | null;
  isFlagged?: boolean;
  createdAt: string;
  updatedAt: string;
}

export type ArgumentInputStance = typeof ArgumentInputStance[keyof typeof ArgumentInputStance];


export const ArgumentInputStance = {
  for: 'for',
  against: 'against',
  neutral: 'neutral',
} as const;

export interface ArgumentInput {
  /** @minLength 10 */
  content: string;
  stance: ArgumentInputStance;
}

export interface ArgumentUpdate {
  /** @minLength 10 */
  content?: string;
}

export interface Reply {
  id: number;
  content: string;
  argumentId: number;
  authorId: number;
  author: User;
  upvotes: number;
  createdAt: string;
}

export interface ReplyInput {
  /** @minLength 1 */
  content: string;
}

export type VoteInputVote = typeof VoteInputVote[keyof typeof VoteInputVote];


export const VoteInputVote = {
  up: 'up',
  down: 'down',
  none: 'none',
} as const;

export interface VoteInput {
  vote: VoteInputVote;
}

export interface VoteResult {
  upvotes: number;
  downvotes: number;
  /** @nullable */
  userVote: string | null;
}

export interface AnalyticsDashboard {
  totalDebates: number;
  activeDebates: number;
  totalArguments: number;
  totalUsers: number;
  totalVotes: number;
  mostActiveDebates: Debate[];
  controversialTopics: Debate[];
}

export interface TrendingCategory {
  name: string;
  debateCount: number;
  argumentCount: number;
  engagementScore: number;
}

export interface LeaderboardEntry {
  rank: number;
  user: User;
  reputation: number;
  tier: string;
  debateCount: number;
  argumentCount: number;
}

export interface ActivityPoint {
  date: string;
  debates: number;
  arguments: number;
  votes: number;
}

export type ListDebatesParams = {
category?: string;
search?: string;
sort?: ListDebatesSort;
limit?: number;
offset?: number;
};

export type ListDebatesSort = typeof ListDebatesSort[keyof typeof ListDebatesSort];


export const ListDebatesSort = {
  latest: 'latest',
  trending: 'trending',
  most_active: 'most_active',
} as const;

export type GetTrendingDebatesParams = {
limit?: number;
};

export type GetLeaderboardParams = {
limit?: number;
};