# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DebateSphere API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /debates/{id}/arguments:
    get:
      operationId: listArguments
      tags: [arguments]
      summary: List arguments for a debate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Arguments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Argument"
    post:
      operationId: createArgument
      tags: [arguments]
      summary: Post an argument
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArgumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Argument"
  /arguments/{id}:
    patch:
      operationId: updateArgument
      tags: [arguments]
      summary: Update an argument
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArgumentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Argument"
    delete:
      operationId: deleteArgument
      tags: [arguments]
      summary: Delete an argument
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /arguments/{id}/replies:
    get:
      operationId: listReplies
      tags: [arguments]
      summary: List replies to an argument
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Replies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reply"
    post:
      operationId: createReply
      tags: [arguments]
      summary: Reply to an argument
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reply"
  /votes/argument/{id}:
    post:
      operationId: voteArgument
      tags: [votes]
      summary: Vote on an argument
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteInput"
      responses:
        "200":
          description: Vote recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteResult"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, username, email, role, reputation, tier, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]
        reputation:
          type: integer
        tier:
          type: string
          enum: [Bronze Debater, Silver Debater, Gold Debater, Diamond Debater]
        bio:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Argument:
      type: object
      required: [id, content, debateId, authorId, author, upvotes, downvotes, replyCount, stance, createdAt, updatedAt]
      properties:
        id:
          type: integer
        content:
          type: string
        debateId:
          type: integer
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/User"
        upvotes:
          type: integer
        downvotes:
          type: integer
        replyCount:
          type: integer
        stance:
          type: string
          enum: [for, against, neutral]
        userVote:
          type: ["string", "null"]
        isFlagged:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ArgumentInput:
      type: object
      required: [content, stance]
      properties:
        content:
          type: string
          minLength: 10
        stance:
          type: string
          enum: [for, against, neutral]
    ArgumentUpdate:
      type: object
      properties:
        content:
          type: string
          minLength: 10
    Reply:
      type: object
      required: [id, content, argumentId, authorId, author, upvotes, createdAt]
      properties:
        id:
          type: integer
        content:
          type: string
        argumentId:
          type: integer
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/User"
        upvotes:
          type: integer
        createdAt:
          type: string
          format: date-time
    ReplyInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
    VoteInput:
      type: object
      required: [vote]
      properties:
        vote:
          type: string
          enum: [up, down, none]
    VoteResult:
      type: object
      required: [upvotes, downvotes, userVote]
      properties:
        upvotes:
          type: integer
        downvotes:
          type: integer
        userVote:
          type: ["string", "null"]
...