# lib/api-zod/src/generated/types/argument.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DebateSphere API
 * OpenAPI spec version: 0.1.0
 */
import type { ArgumentStance } from './argumentStance';
import type { User } from './user';

export interface Argument {
  id: number;
  content: string;
  debateId: number;
  authorId: number;
  author: User;
  upvotes: number;
  downvotes: number;
  replyCount: number;
  stance: ArgumentStance;
  /** @nullable */
  userVote?: string | null;
  isFlagged?: boolean;
  createdAt: Date;
  updatedAt: Date;
}