# lib/api-zod/src/generated/types/argumentUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DebateSphere API
 * OpenAPI spec version: 0.1.0
 */

export interface ArgumentUpdate {
  /** @minLength 10 */
  content?: string;
}