# lib/api-zod/src/generated/types/reply.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DebateSphere API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface Reply {
  id: number;
  content: string;
  argumentId: number;
  authorId: number;
  author: User;
  upvotes: number;
  createdAt: Date;
}