# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DebateSphere API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserTier } from './userTier';

export interface User {
  id: number;
  username: string;
  email: string;
  role: UserRole;
  reputation: number;
  tier: UserTier;
  /** @nullable */
  bio?: string | null;
  createdAt: Date;
}